requirements.txt
africastalking
python-dotenv
google-generativeai
flask